+++ CHANGES.rst
@@ -22,6 +22,8 @@ Unreleased
 -   Added support for key rotation. A list of keys can be passed as
     ``secret_key``, oldest to newest. The newest key is used for
     signing, all keys are tried for unsigning. :pr:`141`
+-   Removed the default SHA-512 fallback signer from
+    ``default_fallback_signers``. :issue:`155`
 
 
 Version 1.1.0